feat: Add custom object definition context
Introduces a new context to manage custom object definitions, enabling dynamic loading and display of industry-specific data. This includes updates to the AppContext, types, and mock data to support this new functionality.

+++ components/settings/LeadScoringRuleEditModal.tsx
@@ -79,7 +79,7 @@ const LeadScoringRuleEditModal: React.FC<LeadScoringRuleEditModalProps> = ({ isO
                         label="Change score by"
                         type="number"
                         value={formData.points}
-                        onChange={(e) => handleChange('points', parseInt(e.target.value) || 0)}
+                        onChange={(e) => handleChange('points', parseFloat(e.target.value) || 0)}
                         required
                         disabled={isSaving}
                     />
